# docs/spcs-openapi.json
{"openapi":"3.1.0","info":{"title":"Smart Plant Care System API","description":"REST API for SPCS – telemetry & watering commands","version":"1.0.0"},"paths":{"/api/token":{"post":{"summary":"Login For Access Token","operationId":"login_for_access_token_api_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"summary":"Health Check","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/telemetry/latest":{"get":{"summary":"Get Latest Telemetry","operationId":"get_latest_telemetry_api_telemetry_latest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/water":{"post":{"summary":"Send Watering Command","operationId":"send_watering_command_api_water_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WateringRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/analysis/status":{"get":{"summary":"Get Plant Status","operationId":"get_plant_status_api_analysis_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlantStatus"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_for_access_token_api_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_token_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PlantStatus":{"properties":{"soil_moisture":{"type":"number","title":"Soil Moisture"},"temperature":{"type":"number","title":"Temperature"},"light_level":{"type":"number","title":"Light Level"},"status":{"type":"string","title":"Status"},"advice":{"type":"string","title":"Advice"}},"type":"object","required":["soil_moisture","temperature","light_level","status","advice"],"title":"PlantStatus"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WateringRequest":{"properties":{"duration_seconds":{"type":"integer","title":"Duration Seconds"}},"type":"object","required":["duration_seconds"],"title":"WateringRequest"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/token"}}}}}}